requirement.txt
asgiref==3.5.2
Django==4.1
django-shopping-cart==0.1
Pillow==9.2.0
psycopg2==2.9.3
sqlparse==0.4.2
tzdata==2022.1
